+++ src/types/mutations/mutation.ts
@@ -2,6 +2,7 @@ import { GraphQLObjectType } from "graphql";
 import { setField } from "./set";
 import { Context } from "../../types";
 import { addWorkspaceField } from "./addWorkspace";
+import { removeWorkspaceField } from "./removeWorkspace";
 import { syncWithPubField } from "./syncWithPub";
 import { ingestDocumentsField } from "./ingestDocuments";
 import { syncManyField } from "./syncMany";
@@ -12,6 +13,7 @@ export const mutationType = new GraphQLObjectType<{}, Context>({
   fields: () => ({
     set: setField,
     addWorkspace: addWorkspaceField,
+    removeWorkspace: removeWorkspaceField,
     syncWithPub: syncWithPubField,
     syncMany: syncManyField,
     ingestDocuments: ingestDocumentsField,