+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features
 
+- Added a `removeWorkspace` mutation which removes a workspace with a given address from the context. It can be authorised using a  `canRemoveWorkspace` checking function provided to the schema.
 - When a workspace with an invalid address is created during the `addWorkspace`, `syncWithPub`, or `syncMany` mutations, the error will be gracefully caught and made part of the expected GraphQL response.
 
 ### Fixes